+++ vllm/model_executor/layers/fused_moe/oracle/fp8.py
@@ -145,7 +145,7 @@ def backend_to_kernel_cls(
         return [MarlinExperts]
 
     elif backend == Fp8MoeBackend.TRITON:
-        from vllm.model_executor.layers.fused_moe.fused_moe import (
+        from vllm.model_executor.layers.fused_moe.experts.triton_moe import (
             TritonExperts,
         )
